Integrate monitoring stubs
Currently does not consider any definitions as monitored, and so should be a
nop, however the structure of the monitor generation is established. The
monitors are generated during the semantic pass, doing the symbol replacement
when definitions are initially gathered, and then generating the monitors after
the underlying definition has been typed checked, based on its type.