Correctly replace the symbols to point to the monitor
The crux of the problem turned out to be that TypeProgram is stateful, storing
the associated LLVM name, and TypeProgram::getCopy() also copies this
state. Also, there was one place in semantic analysis where the name directly
from the DefineProgramContext was used instead of from the associated symbol.

In addition, as part of testing this, the monitor is given access to abort() to
exit.